Autocomplete: extract shared logic into a separate folder
