+++ vscode/src/completions/fast-path-client.ts
@@ -25,7 +25,7 @@ import { SpanStatusCode } from '@opentelemetry/api'
 import type { CompletionResponseWithMetaData } from '@sourcegraph/cody-shared/src/inferenceClient/misc'
 import { logDebug } from '../log'
 import { createRateLimitErrorFromResponse } from './default-client'
-import type { GenerateCompletionsOptions } from './providers/provider'
+import type { GenerateCompletionsOptions } from './providers/shared/provider'
 
 interface FastPathParams extends Pick<GenerateCompletionsOptions, 'authStatus'> {
     isLocalInstance: boolean
