+++ vscode/src/completions/providers/anthropic.ts
@@ -31,13 +31,13 @@ import {
 import {
     type FetchCompletionResult,
     fetchAndProcessDynamicMultilineCompletions,
-} from './fetch-and-process-completions'
+} from './shared/fetch-and-process-completions'
 import {
     type CompletionProviderTracer,
     type GenerateCompletionsOptions,
     Provider,
     type ProviderFactoryParams,
-} from './provider'
+} from './shared/provider'
 
 let isOutdatedSourcegraphInstanceWithoutAnthropicAllowlist = false
