+++ vscode/src/completions/providers/openaicompatible.ts
@@ -11,13 +11,13 @@ import { logDebug } from '../../log'
 import {
     type FetchCompletionResult,
     fetchAndProcessDynamicMultilineCompletions,
-} from './fetch-and-process-completions'
+} from './shared/fetch-and-process-completions'
 import {
     type CompletionProviderTracer,
     type GenerateCompletionsOptions,
     Provider,
     type ProviderFactoryParams,
-} from './provider'
+} from './shared/provider'
 
 class OpenAICompatibleProvider extends Provider {
     public async generateCompletions(
