+++ vscode/src/completions/providers/shared/__mocks__/create-provider-mocks.ts
