+++ vscode/src/completions/providers/shared/dynamic-multiline.ts
@@ -1,7 +1,8 @@
 import type { DocumentContext } from '@sourcegraph/cody-shared'
-import { addAutocompleteDebugEvent } from '../../services/open-telemetry/debug-utils'
-import { insertIntoDocContext } from '../get-current-doc-context'
-import { getFirstLine } from '../text-processing'
+
+import { addAutocompleteDebugEvent } from '../../../services/open-telemetry/debug-utils'
+import { insertIntoDocContext } from '../../get-current-doc-context'
+import { getFirstLine } from '../../text-processing'
 
 interface GetUpdatedDocumentContextParams {
     insertText: string
