+++ vscode/src/completions/providers/shared/parse-provider-and-model.ts
